requirements.txt
numpy==2.0.2
pandas==2.2.3
pickleshare==0.7.5
Flask==3.1.0
python-dotenv==1.0.1